+++ src/join/composer.jl
@@ -519,7 +519,7 @@ innerjoin(df1::AbstractDataFrame, df2::AbstractDataFrame, dfs::AbstractDataFrame
               matchmissing=matchmissing)
 
 """
-    leftjoin(df1, df2; on, makeunique=false, indicator=nothing, validate=(false, false),
+    leftjoin(df1, df2; on, makeunique=false, source=nothing, validate=(false, false),
              renamecols=(identity => identity), matchmissing=:error)
 
 Perform a left join of twodata frame objects and return a `DataFrame` containing
@@ -544,7 +544,7 @@ change in future releases.
   if duplicate names are found in columns not joined on;
   if `true`, duplicate names will be suffixed with `_i`
   (`i` starting at 1 for the first duplicate).
-- `indicator` : Default: `nothing`. If a `Symbol` or string, adds categorical indicator
+- `source` : Default: `nothing`. If a `Symbol` or string, adds indicator
   column with the given name, for whether a row appeared in only `df1` (`"left_only"`),
   only `df2` (`"right_only"`) or in both (`"both"`). If the name is already in use,
   the column name will be modified if `makeunique=true`.
@@ -636,22 +636,35 @@ julia> leftjoin(name, job2, on = [:ID => :identifier], renamecols = uppercase =>
 ```
 """
 function leftjoin(df1::AbstractDataFrame, df2::AbstractDataFrame;
-         on::Union{<:OnType, AbstractVector} = Symbol[],
-         makeunique::Bool=false, indicator::Union{Nothing, Symbol, AbstractString} = nothing,
+         on::Union{<:OnType, AbstractVector} = Symbol[], makeunique::Bool=false,
+         source::Union{Nothing, Symbol, AbstractString}=nothing,
+         indicator::Union{Nothing, Symbol, AbstractString}=nothing,
          validate::Union{Pair{Bool, Bool}, Tuple{Bool, Bool}}=(false, false),
          renamecols::Pair=identity => identity, matchmissing::Symbol=:error)
     if !all(x -> x isa Union{Function, AbstractString, Symbol}, renamecols)
         throw(ArgumentError("renamecols keyword argument must be a `Pair` " *
                             "containing functions, strings, or `Symbol`s"))
     end
+    if source === nothing
+        if indicator !== nothing
+            source = indicator
+            Base.depwarn("`indicator` keyword argument is deprecated and " *
+                         "will be removed in 2.0 release of DataFrames.jl. " *
+                         "Use `source` keyword argument instead.", :leftjoin)
+        end
+    elseif indicator !== nothing
+        throw(ArgumentError("`indicator` keyword argument is deprecated. " *
+                            "It is not allowed to pass both `indicator` and `source` " *
+                            "keyword arguments at the same time."))
+    end
     return _join(df1, df2, on=on, kind=:left, makeunique=makeunique,
-                 indicator=indicator, validate=validate,
+                 indicator=source, validate=validate,
                  left_rename=first(renamecols), right_rename=last(renamecols),
                  matchmissing=matchmissing)
 end
 
 """
-    rightjoin(df1, df2; on, makeunique=false, indicator = nothing,
+    rightjoin(df1, df2; on, makeunique=false, source=nothing,
               validate=(false, false), renamecols=(identity => identity),
               matchmissing=:error)
 
@@ -677,7 +690,7 @@ change in future releases.
   if duplicate names are found in columns not joined on;
   if `true`, duplicate names will be suffixed with `_i`
   (`i` starting at 1 for the first duplicate).
-- `indicator` : Default: `nothing`. If a `Symbol` or string, adds categorical indicator
+- `source` : Default: `nothing`. If a `Symbol` or string, adds indicator
   column with the given name for whether a row appeared in only `df1` (`"left_only"`),
   only `df2` (`"right_only"`) or in both (`"both"`). If the name is already in use,
   the column name will be modified if `makeunique=true`.
@@ -770,21 +783,34 @@ julia> rightjoin(name, job2, on = [:ID => :identifier], renamecols = uppercase =
 """
 function rightjoin(df1::AbstractDataFrame, df2::AbstractDataFrame;
           on::Union{<:OnType, AbstractVector} = Symbol[], makeunique::Bool=false,
-          indicator::Union{Nothing, Symbol, AbstractString} = nothing,
+          source::Union{Nothing, Symbol, AbstractString}=nothing,
+          indicator::Union{Nothing, Symbol, AbstractString}=nothing,
           validate::Union{Pair{Bool, Bool}, Tuple{Bool, Bool}}=(false, false),
           renamecols::Pair=identity => identity, matchmissing::Symbol=:error)
     if !all(x -> x isa Union{Function, AbstractString, Symbol}, renamecols)
         throw(ArgumentError("renamecols keyword argument must be a `Pair` " *
                             "containing functions, strings, or `Symbol`s"))
     end
+    if source === nothing
+        if indicator !== nothing
+            source = indicator
+            Base.depwarn("`indicator` keyword argument is deprecated and " *
+                         "will be removed in 2.0 release of DataFrames.jl. " *
+                         "Use `source` keyword argument instead.", :rightjoin)
+        end
+    elseif indicator !== nothing
+        throw(ArgumentError("`indicator` keyword argument is deprecated. " *
+                            "It is not allowed to pass both `indicator` and `source` " *
+                            "keyword arguments at the same time."))
+    end
     return _join(df1, df2, on=on, kind=:right, makeunique=makeunique,
-                 indicator=indicator, validate=validate,
+                 indicator=source, validate=validate,
                  left_rename=first(renamecols), right_rename=last(renamecols),
                  matchmissing=matchmissing)
 end
 
 """
-    outerjoin(df1, df2; on, makeunique=false, indicator=nothing, validate=(false, false),
+    outerjoin(df1, df2; on, makeunique=false, source=nothing, validate=(false, false),
               renamecols=(identity => identity), matchmissing=:error)
     outerjoin(df1, df2, dfs...; on, makeunique = false,
               validate = (false, false), matchmissing=:error)
@@ -814,7 +840,7 @@ This behavior may change in future releases.
   if duplicate names are found in columns not joined on;
   if `true`, duplicate names will be suffixed with `_i`
   (`i` starting at 1 for the first duplicate).
-- `indicator` : Default: `nothing`. If a `Symbol` or string, adds categorical indicator
+- `source` : Default: `nothing`. If a `Symbol` or string, adds indicator
   column with the given name for whether a row appeared in only `df1` (`"left_only"`),
   only `df2` (`"right_only"`) or in both (`"both"`). If the name is already in use,
   the column name will be modified if `makeunique=true`.
@@ -913,16 +939,29 @@ julia> rightjoin(name, job2, on = [:ID => :identifier], renamecols = uppercase =
 ```
 """
 function outerjoin(df1::AbstractDataFrame, df2::AbstractDataFrame;
-          on::Union{<:OnType, AbstractVector} = Symbol[],
-          makeunique::Bool=false, indicator::Union{Nothing, Symbol, AbstractString} = nothing,
+          on::Union{<:OnType, AbstractVector} = Symbol[], makeunique::Bool=false,
+          source::Union{Nothing, Symbol, AbstractString}=nothing,
+          indicator::Union{Nothing, Symbol, AbstractString}=nothing,
           validate::Union{Pair{Bool, Bool}, Tuple{Bool, Bool}}=(false, false),
           renamecols::Pair=identity => identity, matchmissing::Symbol=:error)
     if !all(x -> x isa Union{Function, AbstractString, Symbol}, renamecols)
         throw(ArgumentError("renamecols keyword argument must be a `Pair` " *
                             "containing functions, strings, or `Symbol`s"))
     end
+    if source === nothing
+        if indicator !== nothing
+            source = indicator
+            Base.depwarn("`indicator` keyword argument is deprecated and " *
+                         "will be removed in 2.0 release of DataFrames.jl. " *
+                         "Use `source` keyword argument instead.", :outerjoin)
+        end
+    elseif indicator !== nothing
+        throw(ArgumentError("`indicator` keyword argument is deprecated. " *
+                            "It is not allowed to pass both `indicator` and `source` " *
+                            "keyword arguments at the same time."))
+    end
     return _join(df1, df2, on=on, kind=:outer, makeunique=makeunique,
-                 indicator=indicator, validate=validate,
+                 indicator=source, validate=validate,
                  left_rename=first(renamecols), right_rename=last(renamecols),
                  matchmissing=matchmissing)
 end